#22 Create reducers to filter top entries (rough)

+++ src/ch/ljacqu/wordeval/evaluation/export/ExportParams.java
@@ -8,7 +8,7 @@ public class ExportParams {
    * If not null, keys must be bigger or equals to <code>minimum</code> to be
    * kept in full form.
    */
-  public final Integer minimum;
+  public final Double minimum;
   /** Whether or not to show the results in descending order. */
   public final boolean isDescending;
   /** The maximum number of entries per key to keep. */
@@ -16,7 +16,7 @@ public class ExportParams {
   /** For PartWordExport: the maximum size a word list may have. */
   public final Integer maxPartWordListSize;
 
-  ExportParams(int number, Integer minimum, boolean isDescending,
+  ExportParams(int number, Double minimum, boolean isDescending,
       Integer maxEntry, Integer maxPartWordListSize) {
     this.topKeys = number;
     this.minimum = minimum;
