+++ src/ch/ljacqu/wordeval/evaluation/export/ExportParamsBuilder.java
@@ -3,7 +3,7 @@ package ch.ljacqu.wordeval.evaluation.export;
 public class ExportParamsBuilder {
 
   private int topKeys = 5;
-  private Integer minimum = null;
+  private Double minimum = null;
   private boolean isDescending = true;
   private Integer maxTopEntrySize = 50;
   private Integer maxPartWordListSize = null;
@@ -18,7 +18,7 @@ public class ExportParamsBuilder {
     return this;
   }
 
-  public ExportParamsBuilder setMinimum(Integer minimum) {
+  public ExportParamsBuilder setMinimum(Double minimum) {
     this.minimum = minimum;
     return this;
   }
